Fix absolute imports with ttsc and transform plugin
Background to why this is needed: https://medium.com/@grrowl/fixing-absolute-imports-in-typescript-797f405176eb
In short: with baseUrl you can use absolute imports in code.
This makes the imports less reliant on the relative code dir structure.
However TS doesn't support transforming those absolute imports when the code is
compiled to JS. This leaves require('components/Button') imports to the JS code
which doesn't work. The solution for this is ttsc, a wrapper for tsc which allows
defining transformation plugins in tsconfig.json. We are then using the
@zerollup/ts-transform-paths plugin to do the actual transformation.